Ejercicios Bases de datos 2, 2 B, 3 y 3 B: añado $registroNoVacioOk para comprobación registro no es vacío (que ya hacía pero sin variable Ok)

+++ ejercicios/bases-datos/bases-de-datos-2-b/iaw-agenda/borrar-2.php
@@ -13,10 +13,18 @@ cabecera("Borrar 2", MENU_VOLVER);
 
 $id = recoge("id", []);
 
+$idOk = false;
+
 if (count($id) == 0) {
     print "    <p class=\"aviso\">No se ha seleccionado ningún registro.</p>\n";
 } else {
+    $idOk = true;
+}
+
+if ($idOk) {
     foreach ($id as $indice => $valor) {
+        $registroEncontradoOk = false;
+
         $consulta = "SELECT COUNT(*) FROM $cfg[tablaPersonas]
                      WHERE id = :indice";
 
@@ -28,6 +36,10 @@ if (count($id) == 0) {
         } elseif ($resultado->fetchColumn() == 0) {
             print "    <p class=\"aviso\">Registro no encontrado.</p>\n";
         } else {
+            $registroEncontradoOk = true;
+        }
+
+        if ($registroEncontradoOk) {
             $consulta = "DELETE FROM $cfg[tablaPersonas]
                          WHERE id = :indice";
