+++ ejercicios/bases-datos/bases-de-datos-3-c/bases-de-datos-3-b-0/db/tabla-usuarios/borrar-2.php
@@ -21,11 +21,19 @@ cabecera("Usuarios - Borrar 2", MENU_USUARIOS, PROFUNDIDAD_2);
 
 $id = recoge("id", []);
 
+$idOk = false;
+
 if (count($id) == 0) {
     print "    <p class=\"aviso\">No se ha seleccionado ningún registro.</p>\n";
 } else {
+    $idOk = true;
+}
+
+if ($idOk) {
     foreach ($id as $indice => $valor) {
-        $consulta = "SELECT * FROM $cfg[tablaUsuarios]
+        $registroEncontradoOk = false;
+
+        $consulta = "SELECT COUNT(*) FROM $cfg[tablaUsuarios]
                      WHERE id = :indice";
 
         $resultado = $pdo->prepare($consulta);
@@ -33,11 +41,32 @@ if (count($id) == 0) {
             print "    <p class=\"aviso\">Error al preparar la consulta. SQLSTATE[{$pdo->errorCode()}]: {$pdo->errorInfo()[2]}</p>\n";
         } elseif (!$resultado->execute([":indice" => $indice])) {
             print "    <p class=\"aviso\">Error al ejecutar la consulta. SQLSTATE[{$pdo->errorCode()}]: {$pdo->errorInfo()[2]}</p>\n";
-        } elseif (!($registro = $resultado->fetch())) {
+        } elseif ($resultado->fetchColumn() == 0) {
             print "    <p class=\"aviso\">Registro no encontrado.</p>\n";
-        } elseif ($registro["usuario"] == $cfg["rootName"]) {
-            print "    <p class=\"aviso\">Este usuario no se puede borrar.</p>\n";
         } else {
+            $registroEncontradoOk = true;
+        }
+
+        $registroNoRootOk = false;
+
+        if ($registroEncontradoOk) {
+            $consulta = "SELECT COUNT(*) FROM $cfg[tablaUsuarios]
+                         WHERE id = :indice
+                         AND usuario = :usuario";
+
+            $resultado = $pdo->prepare($consulta);
+            if (!$resultado) {
+                print "    <p class=\"aviso\">Error al preparar la consulta. SQLSTATE[{$pdo->errorCode()}]: {$pdo->errorInfo()[2]}</p>\n";
+            } elseif (!$resultado->execute([":indice" => $indice, ":usuario" => $cfg["rootName"]])) {
+                print "    <p class=\"aviso\">Error al ejecutar la consulta. SQLSTATE[{$pdo->errorCode()}]: {$pdo->errorInfo()[2]}</p>\n";
+            } elseif ($resultado->fetchColumn() > 0) {
+                print "    <p class=\"aviso\">Este usuario no se puede borrar.</p>\n";
+            } else {
+                $registroNoRootOk = true;
+            }
+        }
+
+        if ($registroEncontradoOk && $registroNoRootOk) {
             $consulta = "DELETE FROM $cfg[tablaUsuarios]
                          WHERE id = :indice";
